+++ src/core/ports/index.ts
@@ -1,4 +1,5 @@
-export * from './todo.repository';
 export * from './jsonplaceholder.integration';
 export * from './todo.cache';
 export * from './todo.producer';
+export * from './todo.collection';
+export * from './todo.repository';