+++ src/main/modules/index.ts
@@ -3,3 +3,4 @@ export * from './cli/cli';
 export * from './http/http';
 export * from './amqp/amqp';
 export * from './graphql/graphql';
+export * from './mongodb/mongodb';